Scripts to generate CISM grids
The folder libs contains functions and constants used throughout CISM
pre- and post-processing tools.

The folder Grids contains tools allowing for the creation of CISM
standalone grids and their lat-lon equivalent. It also contains script
and instruction on how to generate a SCRIP file and an ESMF mesh file
(to be used for CESM mapping files).